final project submission for Machine Learning project university of Milan 2023
Neural Networks
Used Pytorch to train a neural network for the binary classification of muffins and Chihuahuas based on images from this dataset. 
Images transformed from JPG to RGB pixel values and scaled down. 
The student is asked to:
experiment with different network architectures (at least 3) and training hyperparameters,
use 5-fold cross-validation to compute your risk estimates,
thoroughly discuss the obtained results, documenting the influence of the choice of the network architecture and the tuning of the hyperparameters on the final cross-validated risk estimate.
While the training loss can be chosen freely, the reported cross-validated estimates must be computed according to the zero-one loss.